Update scraper urls now that there is a new term
The url in the scraper was not working anymore, so it was loading
the last term page by default, but assigning the data to the previous
term through a hardcoded term field.
This change updates the scraper with both urls (to the previous term
and the new term) and assigns the incoming data from those urls to
the right terms.